+++ certbot/client.py
@@ -27,7 +27,6 @@ from certbot import eff
 from certbot import error_handler
 from certbot import errors
 from certbot import interfaces
-from certbot import reverter
 from certbot import storage
 from certbot import util
 from certbot.compat import os
@@ -709,20 +708,6 @@ def rollback(default_installer, checkpoints, config, plugins):
     if installer is not None:
         installer.rollback_checkpoints(checkpoints)
         installer.restart()
-
-
-def view_config_changes(config):
-    """View checkpoints and associated configuration changes.
-
-    .. note:: This assumes that the installation is using a Reverter object.
-
-    :param config: Configuration.
-    :type config: :class:`certbot.interfaces.IConfig`
-
-    """
-    rev = reverter.Reverter(config)
-    rev.recovery_routine()
-    rev.view_config_changes()
 
 def _open_pem_file(cli_arg_path, pem_path):
     """Open a pem file.